Move useIsFirstRender() to hooks/index.js
