+++ src/utils/index.js
@@ -1,4 +1,3 @@
-import { useEffect, useState } from 'react';
 import slugify from 'slugify';
 
 // so it doesn't throw if no window
@@ -66,19 +65,4 @@ export const shuffleCopy = (array) => {
     copy[j] = temp;
   }
   return copy;
-};
-
-/**
- * This hook will return true for the first render on the server and for the
- * first render on the client. Otherwhise, it will return false.
- *
- * This function encapsulates a hook so "rules of hooks" apply to it.
- * @see {@link https://reactjs.org/docs/hooks-rules.html}
- *
- * @return {boolean} true on the first server side and client side render
- */
-export const useIsFirstRender = () => {
-  const [isFirst, setIsFirst] = useState(true);
-  useEffect(() => setIsFirst(false), []);
-  return isFirst;
 };
